feat: set up side nav prefetch and change stale time to 10 min
