+++ src/context/AppProvider.js
@@ -6,7 +6,7 @@ import { BrowserRouter as Router } from 'react-router-dom'
 const queryClient = new QueryClient({
   defaultOptions: {
     queries: {
-      staleTime: 10000,
+      staleTime: 10 * 60000,
       refetchOnWindowFocus: false,
     },
   },
