--- src/lib/hooks/useBoardsQuery.js
@@ -1,9 +1,0 @@
-import { useQuery } from 'react-query'
-import { fetchBoards } from 'lib/api/fetchers'
-
-export function useBoardsQuery() {
-  return useQuery({
-    queryKey: 'boards',
-    queryFn: fetchBoards,
-  })
-}
